fix: auto stop all remote tracks if muted for long
